Florian Schneider's optimisation patch
Add patch from https://github.com/financeCoding/dart-bignum/issues/8
this is supposed to be an optimisation to make it go faster, it does
so by not using a Map but instead an auto-growing array to do storage
of the internal data structure.

It was also necessary to fix a type error this revealed.

I did not include the additions of parentheses which the patch on github
introduced as they seemed to add unnecessary noise to the patch (and I
don't see what they are for).